glossary_RPCservice/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
SQLAlchemy==2.0.36
alembic==1.13.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
